// web/src/api/generated/models/PlaceDetail.ts
export interface PlaceDetail {
    /**
     * 
     * @type {string}
     * @memberof PlaceDetail
     */
    address: string | null;
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof PlaceDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof PlaceDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof PlaceDetail
     */
    createdBy: string;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof PlaceDetail
     */
    creator: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof PlaceDetail
     */
    description: string | null;
    /**
     * 
     * @type {string}
     * @memberof PlaceDetail
     */
    id: string;
    /**
     * 
     * @type {number}
     * @memberof PlaceDetail
     */
    latitude: number | null;
    /**
     * 
     * @type {number}
     * @memberof PlaceDetail
     */
    longitude: number | null;
    /**
     * 
     * @type {string}
     * @memberof PlaceDetail
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof PlaceDetail
     */
    spaceId: string;
    /**
     * 
     * @type {Date}
     * @memberof PlaceDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof PlaceDetail
     */
    version: number;
}